+++ src/main/java/finalmission/infrastructure/sms/CoolSmsRequest.java
@@ -1,4 +1,4 @@
-package finalmission.infrastructure;
+package finalmission.infrastructure.sms;
 
 import java.util.List;
